requirements.txt
celery==3.1.16
cryptography>=2.5
dnspython
feedparser
lark-parser>=0.7.8
netaddr>=0.7.19
Pillow>=6.0.0
pymongo==3.3.0
pyperclip
pysftp>=0.2.9
python-docx>=0.8.10
python-pptx>=0.6.8
requests>=2.18.4
XlsxWriter>=1.0.2
